requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.1.0
arabic-reshaper==2.1.3
asgiref==3.4.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
autopep8==1.6.0
Babel==2.9.1
billiard==3.6.4.0
bootstrap-admin==0.4.4
Brotli==1.0.9
cached-property==1.5.2
cairocffi==1.0.0
CairoSVG==2.5.2
celery==5.1.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
crispy-bootstrap5==0.6
cryptography==36.0.2
cssselect2==0.4.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.15
django-admin-honeypot==1.1.0
django-allauth==0.50.0
django-bootstrap-modal-forms==2.2.0
django-ckeditor==6.2.0
django-countries==7.2.1
django-crispy-forms==1.13.0
django-currentuser==0.5.3
django-filter==21.1
django-grappelli==2.15.3
django-hijack==3.1.2
django-import-export==2.6.1
django-jazzmin==2.5.0
django-js-asset==1.2.2
django-language-fields==1.0.1
django-material-admin==1.8.5
django-money==2.1
django-phonenumber-field==5.2.0
django-requestlogging==1.0.1
django-rosetta==0.9.8
django-tagify2==1.0.19
django-widget-tweaks==1.4.11
djangorestframework==3.12.4
easy-thumbnails==2.8.1
et-xmlfile==1.1.0
fonttools==4.27.1
frozenlist==1.2.0
future==0.18.2
geoip2==4.6.0
gunicorn==20.1.0
html5lib==1.1
idna==3.3
idna-ssl==1.1.0
importlib-metadata==4.8.3
install==1.3.5
kombu==5.1.0
lxml==4.7.1
Markdown==3.3.4
MarkupPy==1.14
maxminddb==2.2.0
multidict==5.2.0
mysqlclient==2.1.0
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.9
phonenumbers==8.12.36
Pillow==8.4.0
polib==1.1.1
prompt-toolkit==3.0.28
py-moneyed==1.2
pycodestyle==2.8.0
pycparser==2.21
pydyf==0.1.2
PyJWT==2.3.0
PyPDF2==1.26.0
pyphen==0.11.0
python-bidi==0.4.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
reportlab==3.6.5
requests==2.27.1
requests-oauthlib==1.3.1
setproctitle==1.2.2
six==1.16.0
sqlparse==0.4.2
svglib==1.1.0
tablib==3.1.0
tinycss2==1.1.1
toml==0.10.2
typing_extensions==4.1.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
WeasyPrint==52.5
webencodings==0.5.1
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
yarl==1.7.2
zipp==3.6.0
zopfli==0.1.9
